requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
albucore==0.0.24
albumentations==2.0.6
annotated-types==0.7.0
anyio==4.9.0
appscript==1.3.0
astor==0.8.1
attrdict==2.0.1
attrs==25.3.0
babel==2.17.0
bce-python-sdk==0.9.35
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==6.0.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
colorlog==6.9.0
contourpy==1.3.2
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.0.12
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
docx2pdf==0.1.8
einops==0.8.1
et_xmlfile==2.0.0
fastapi==0.115.12
filelock==3.18.0
fire==0.7.0
Flask==3.1.1
flask-babel==4.0.0
fonttools==4.57.0
fpdf2==2.8.3
frozenlist==1.7.0
fsspec==2025.5.1
ftfy==6.3.1
future==1.0.0
GPUtil==1.4.0
h11==0.14.0
hf-xet==1.1.3
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.32.5
idna==3.10
imageio==2.37.0
imagesize==1.4.1
imgaug==0.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.8
langchain==0.2.17
langchain-community==0.2.17
langchain-core==0.2.43
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langsmith==0.1.147
lazy_loader==0.4
lmdb==1.6.2
lxml==5.4.0
markdown2==2.5.3
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
more-itertools==10.7.0
multidict==6.4.4
mypy_extensions==1.1.0
networkx==3.4.2
numpy==1.26.4
openai==1.63.2
opencv-contrib-python==4.10.0.84
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt-einsum==3.3.0
orjson==3.10.18
packaging==24.2
paddleocr==2.6
paddlepaddle==3.0.0
pandas==1.5.3
pdf2image==1.17.0
pillow==11.1.0
premailer==3.10.0
prettytable==3.16.0
propcache==0.3.2
protobuf==6.30.1
psutil==7.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pycryptodome==3.23.0
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
rarfile==4.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.2
setuptools==80.9.0
shapely==2.1.0
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.2
stringzilla==3.12.5
tenacity==8.5.0
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.3.30
tiktoken==0.9.0
tokenizers==0.19.1
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.34.2
visualdl==2.5.3
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.20.1
